Asynchronously handle print events. The previous implementation would allow print events to become buffered up in the TCP channel until an evaluation was done, at which time they would all be drained (and printed in the REPL), followed by the result of the evaluation being printed.
This differs, from, say, the Node.js REPL which inherits the stdout
of the Node.js process, thus having prints occur when they are done.

This buffering could pose a problem if the JSC engine had timers that
involved a continuous stream of print events that were not being
dealt with.

This commit dedicates a thread in the REPL to reading messages off the
input stream. If a print message arrives, it is handled promptly.
On the other hand, if an evaluation response arrives, it is transferred
to the (blocking) -evaluate call via a promise.

Unfortunately, this adds a bit of complexity in the passing of promises
via an atom in the REPL state, and it really adds a lot of complexity
in ensuring that shutdown occurs properly (it has to handle cases so
that the deref of the promise doesn't block forever for a response
that may never arrive). Perhaps one way this complexity could be
addressed would be put pull in core.async as a dependency and set up a
clean CSP approach, but the new implementation in this commit gets by
without being overly difficult to understand.