Set up things so that relative paths are passed for compiler output
It turns out that relative paths work fine, so pushing in this set
of experimental changes. If they fail for some reason, we can choose
another avenue.

The nice thing is that this eliminates messing with the Objective-C
code to get things up and running. For the simulator, to avoid any
config for now, compiler output is simply set to /tmp/ambly-cljs-out.
If we end up using WebDAV for simulator builds as well (which works),
then we could eliminate this difference as well.