Initial cut at auto-configuration via Bonjour.
Some things are still rough around the corners, but it generally works.
The worst issue appears to be a race between when files are written
to WebDAV and them being read in the device/simulator. A large sleep
is in place to try to avoid this race.

The first device/simulator discovered is used. In the future UI should
be added to let the user pick if there are multiple available.