Fix #20 Disable device sleeping
Just unconditionally do this in the demo app.

The startup of the demo app is starting growing a body of code that
could be put behind a useful façade (so that consumers of Ambly can
easily get up and going.) This idle timer stuff could be an option
in such a façade if we have one. But for now, drop it right in the
app.