Move WebDAV into ABYServer as an impl detail. Clean up CocoaPod setup.
To make things more reusable, have the WebDAV stuff hide inside the
ABYServer abstraction. Clean up the CocoaPod setup so that the
GDCWebServer dep is specified as a transitive deep of Ambly.

Further, clean up the Ambly Demo app so that it uses CocoaPods via
a path dependency to pull in Ambly src (and GCDWebServer).