Fix #56 Need to set print fn
For Ambly, the *print-fn* sends a message over the REPL TCP connection
to a connected REPL (using the special terminator 1 to distinguish it
from evaluation responses). Thus the *print-fn* is inextricably tied
to the ABYServer implementation, at least for the purposes of printing
for connected Ambly REPLs.

The *print-fn* cannot be set until after ClojureScript has been
bootstrapped. There are two cases to handle:

- If bootstrapping a raw JSC environment for ClojureScript,
  (as in the Ambly Demo) we do this in the Ambly Clojure REPL code,
  referring to AMBLY_PRINT_FN, which will have been set by ABYServer.

- If bootstrapping an app that has its own bundled JS derived from
  ClojureScript, the bootstrapping occurs on the Objective-C side. In
  this case, when setting up ABYServer, we detect that bootstrapping
  has already occurred and at that point set up *print-fn*.

This commit also cleans up related things like:

- Setting up the the print capability when and only when the server
  successfully listens.

- Making the print capability revert to acting like console.out when
  a REPL is no longer connected.

- Eliminating the `JS` prefix being placed on all console.out writes.

- Avoiding defining setting the print function as out.write, which
  could interfere with something else that might use those symbols, and
  instead going with a distinguished symbol AMBLY_PRINT_FN.