Derive mount point from URL instead of Bonjour name
It is easy for the same endpoint to be reused if
switching between Ambly-enabled apps on a local
dev box. This leads to a situation where the OS can
re-establish the WebDAV mount, and if Bonjour
name is used to form the unique mount directory
name, this can cause a problem when trying to
establish WebDAV mount to new unique directory
name. This resolves this problem by using the
(unchanging) URL to form the mount directory name.

Fixes #86