Instructions turn off delayed upload
If using mount.davfs, the default configuration
uses a delayed upload of 10 seconds, meaning
files written don't appear on the device until
10 seconds after being closed. This causes an
easily lost race when the device attempts to
read a file via `require` or during bootstrapping.

This default delay can be turned off. This change
adds to the mount instructions showing
users how to do this.

Fixes #90.